backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
supabase==2.3.1
Pillow==10.0.1
requests==2.31.0
httpx==0.24.1